Conflicts:
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/dao/ChartDetailDao.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/AccountLog.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/User.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/UserDeviceDetails.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/payment/PSMSPaymentDetails.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/payment/PaymentDetails.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/repository/UserGroupRepository.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/repository/UserRepository.java
	persistence/src/main/resources/META-INF/jpql/chartDetail.xml
	persistence/src/main/resources/sql/release5.0-SNAPSHOT.sql
	persistence/src/test/java/mobi/nowtechnologies/server/persistence/domain/UserTest.java
	persistence/src/test/java/mobi/nowtechnologies/server/persistence/repository/UserRepositoryIT.java
	service/src/main/java/mobi/nowtechnologies/server/service/UserService.java
	service/src/main/java/mobi/nowtechnologies/server/service/impl/UserNotificationServiceImpl.java
	service/src/test/java/mobi/nowtechnologies/server/service/PaymentDetailsServiceTest.java
	service/src/test/java/mobi/nowtechnologies/server/service/UserServiceTest.java
	service/src/test/java/mobi/nowtechnologies/server/service/impl/UserNotificationServiceImplTest.java
	transport/src/main/java/mobi/nowtechnologies/server/job/ForkO2UsersForUpdateJob.java
	transport/src/main/webapp/WEB-INF/conf/log4j.xml

+++ persistence/src/main/java/mobi/nowtechnologies/server/persistence/repository/UserGroupRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.Query;
  * User: Titov Mykhaylo (titov)
  * 09.10.13 15:13
  */
-public interface UserGroupRepository extends JpaRepository<UserGroup, Byte> {
+public interface UserGroupRepository extends JpaRepository<UserGroup, Integer> {
 
     UserGroup findByCommunity(Community community);
